### 2021-05-06 #3261 @whedon
[PRE REVIEW]: flux-data-qaqc: A Python Package for Energy Balance Closure and Post-Processing of Eddy Flux Data
**Submitting author:** @JohnVolk (<a href="http://orcid.org/0000-0001-9994-1545">John Volk</a>)
**Repository:** <a href="https://github.com/Open-ET/flux-data-qaqc" target ="_blank">https://github.com/Open-ET/flux-data-qaqc</a>
**Version:** 0.1.5
