### 2021-06-14 #3261 @pdebuyl
@ashwinvis @rabernat would you review the article "flux-data-qaqc: A Python Package for Energy Balance Closure and Post-Processing of Eddy Flux Data" for the Journal of Open Source Software ?
