### 2021-06-23 #3261 @pdebuyl
@dgketchum would you review the article "flux-data-qaqc: A Python Package for Energy Balance Closure and Post-Processing of Eddy Flux Data" for the Journal of Open Source Software ?
