### 2021-06-25 #3418 @pdebuyl
@ashwinvis @dgketchum make sure to accept the invitation to the reviewers group and to have a look at the reviewer guidelines linked to at the top of this review page.

The review process will happen in this issue page, so questions to the author or to me can be added as comments here. As this is the first JOSS review for both reviewers (unless I missed something), do not hesitate to ask questions if you have doubts about the procedure. You can also take a look at earlier reviews to get an idea of how the reviews proceed https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+label%3Aaccepted+

@JohnVolk both reviewers gave indication that they can't start the review immediately. Given the remaining disruption because of the pandemic, JOSS operates with less strict schedules than usual. As there are also more vacations during the summer, I am already sorry for the delays.
