### 2021-08-30 #3418 @dgketchum
I'm ready to complete the checklist but it is not editable, and the link to the invitation is dead. @pdebuyl, advice?
