### 2021-09-02 b9c6525
add example data and testing components * add all tutorial files to package including Jupyter Notebooks * add .yml configuration file for automated testing with Travis-CI * pytest .ini, configuration, and testing file with single test
