+++ .travis.yml
@@ -1,0 +1,24 @@
+language: python
+
+python:
+  - "3.7"
+
+before_install:
+  - wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh
+  - chmod +x miniconda.sh
+  - ./miniconda.sh -b
+  - export PATH=/home/travis/miniconda2/bin:$PATH
+  - conda update --yes conda
+
+install:
+  - conda env create -f environment.yml
+  - source activate fluxdataqaqc
+  - conda install pytest --yes
+  - conda install pytest-cov --yes
+  - conda install coveralls --yes
+
+script:
+  - python -m pytest -v --cov=fluxdataqaqc
+
+after_success:
+  - coveralls
