### 2021-09-05 607db37
add github actions yml for automated testing
