+++ fluxdataqaqc/util.py
@@ -6,6 +6,7 @@ module.
 
 import numpy as np
 import pandas as pd
+from pathlib import Path
 
 class Convert(object):
     """
