### 2021-09-05 c47c62a
Fix path to conda package in gtihub actions
