+++ .github/workflows/fluxdataqaqc_tests.yml
@@ -20,7 +20,7 @@ jobs:
       run: |
         $CONDA/bin/conda install flake8
         # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics --builtins='Path'
+        $CONDA/bin/flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics --builtins='Path'
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         $CONDA/bin/flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
