+++ requirements.txt
@@ -12,7 +12,7 @@ olefile==0.46
 openpyxl==3.0.5
 packaging==20.4
 pandas==1.1.4
-Pillow==8.2.0
+Pillow==8.3.2
 pip==20.2.4
 pyparsing==2.4.7
 python-dateutil==2.8.1
