### 2021-09-19 c9db99c
* Add pytest automated tests for Data class * some tests for Convert and Plot classes * bug fix in Data.plot method
