+++ tests/test_fluxdataqaqc.py
@@ -54,12 +54,122 @@ def data(request):
     
 class TestData(object):
 
-#    def setup_method(self, data):
-#        self.root_dir = data['package_root_dir']
-        
-    def test_ACSE_refET(self, data):
-        config = data['package_root_dir']/'examples'/'Basic_usage'/'US-Tw3_config.ini'
-        data_obj = Data(config)
-        ts = data_obj.hourly_ASCE_refET()
+    @pytest.fixture(autouse=True)
+    def setup_method(self, data):
+        config = data['package_root_dir']\
+            /'examples'/'Basic_usage'/'US-Tw3_config.ini'
+        # for testing auto QC flag filtering
+        self.fluxnet_config = data['package_root_dir']\
+            /'examples'/'Basic_usage'/'fluxnet_config.ini'
+        self.data_obj = Data(config)
+        self.header_Tw3 = [
+            'TIMESTAMP_START', 'TIMESTAMP_END', 'CO2', 'H2O', 'CH4', 'FC',
+            'FCH4', 'FC_SSITC_TEST', 'FCH4_SSITC_TEST', 'G', 'H', 'LE',
+            'H_SSITC_TEST', 'LE_SSITC_TEST', 'WD', 'WS', 'USTAR', 'ZL', 'TAU',
+            'MO_LENGTH', 'V_SIGMA', 'W_SIGMA', 'TAU_SSITC_TEST', 'PA', 'RH',
+            'TA', 'VPD_PI', 'T_SONIC', 'T_SONIC_SIGMA', 'SWC_1_1_1',
+            'SWC_1_2_1', 'TS_1_1_1', 'TS_1_2_1', 'TS_1_3_1', 'TS_1_4_1',
+            'TS_1_5_1', 'NETRAD', 'PPFD_DIF', 'PPFD_IN', 'PPFD_OUT', 'SW_IN',
+            'SW_OUT', 'LW_IN', 'LW_OUT', 'P', 'FC_PI_F', 'RECO_PI_F',
+            'GPP_PI_F', 'H_PI_F', 'LE_PI_F'
+        ]
+
+
+    def test_Data_init(self):
+        # checking constructor of Data class
+        assert self.data_obj.climate_file.name == 'AMF_US-Tw3_BASE_HH_5-5.csv'
+        assert self.data_obj.config_file.stem == 'US-Tw3_config'
+        assert self.data_obj.elevation == -9
+        assert self.data_obj.latitude == 38.1159
+        # dataframe should not be loaded
+        assert self.data_obj._df == None
+        # header of climate file should have been read however
+        assert (self.data_obj.header == self.header_Tw3).all()
+        assert self.data_obj.units.get('wd') == 'azimuth (degrees)'
+
+
+    def test_config_parser(self):
+        assert self.data_obj.config.sections() == ['METADATA', 'DATA']
+        assert self.data_obj.config.get('METADATA', 'igbp') == 'CRO'
+
+    def test_units_and_Convert_inheritance(self):
+        units = self.data_obj.units
+        assert isinstance(units, dict)
+        assert units.get('LE') == 'w/m2'
+        assert self.data_obj.pretty_unit_names.get('kpa') == 'kPa'
+
+    def test_Data_load_df(self):
+        # check on loading climate timeseries input data into Pandas DataFrame
+        df = self.data_obj.df
+        assert isinstance(df, pd.DataFrame)
+        assert len(df.dropna()) == 68035
+
+    def test_datetime_index(self):
+        df = self.data_obj.df
+        assert isinstance(df.index[0], pd.Timestamp)
+        assert df.index[0].year == 2013
+
+    def test_auto_average(self):
+        # check that automatic averaging of soil moisture was done
+        df = self.data_obj.df
+        assert 'theta_mean' in df.columns
[…]
