+++ pytest.ini
@@ -2,3 +2,4 @@
 filterwarnings =
     ignore::DeprecationWarning
     ignore::RuntimeWarning
+    ignore::PendingDeprecationWarning
