### 2021-09-19 5e28aaf
remove xlrd reader to fix incompatibility with new excel versions
