### 2021-09-19 cbfccc4
remove xlrd test and all name arguments from bokeh figures for testing error
