+++ fluxdataqaqc/plot.py
@@ -199,7 +199,6 @@ class Plot(object):
             This method is also available from the :obj:`.Data` and :obj:`.QaQc`
             objects.
         """
-        name = '{}_vs_{}'.format(x,y)
         # remove pairs where one or both are nans, for LSRL and min/max-1:1
         
         if not y in source.data:
@@ -432,7 +431,7 @@ class Plot(object):
         y_label = _get_units(plt_vars, units)
         fig = figure(
             x_axis_label=x_label, y_axis_label=y_label, title=title,
-            width=plot_width, height=plot_height, name='energy_balance_daily'
+            width=plot_width, height=plot_height, 
         )
         fig = Plot.add_lines(
             fig, df, plt_vars, colors, x_label, daily_source, labels=plt_vars
@@ -449,7 +448,6 @@ class Plot(object):
             title = 'Monthly Surface Energy Balance Components'
             fig = figure(x_axis_label=x_label, y_axis_label=y_label,title=title,
                 width=plot_width, height=plot_height, 
-                name='energy_balance_monthly'
             )
             fig = Plot.add_lines(
                 fig, monthly_df, plt_vars, colors, x_label, monthly_source,
@@ -472,7 +470,6 @@ class Plot(object):
             y_label = _get_units(plt_vars, units)
             fig = figure(x_axis_label=x_label, y_axis_label=y_label, 
                 title=title, width=plot_width, height=plot_height, 
-                name='Rs_daily'
             )
             fig = Plot.add_lines(
                 fig, df, plt_vars, colors, x_label, daily_source, labels=labels
@@ -516,7 +513,7 @@ class Plot(object):
             y_label = _get_units(g_vars, units)
             fig = figure(
                 x_axis_label=x_label, y_axis_label=y_label, title=title,
-                plot_width=plot_width, plot_height=plot_height, name='G_daily'
+                plot_width=plot_width, plot_height=plot_height
             )
             fig = Plot.add_lines(
                 fig, tmp_df, plt_vars, colors, x_label, tmp_source, 
@@ -538,7 +535,6 @@ class Plot(object):
                     fig = figure(
                         x_axis_label=x_label, y_axis_label=y_label,title=title,
                         plot_width=plot_width, plot_height=plot_height, 
-                        name='G_monthly'
                     )
                     fig = Plot.add_lines(
                         fig, tmp_df, plt_vars, colors, x_label, tmp_source,
@@ -557,8 +553,7 @@ class Plot(object):
         y_label = _get_units(plt_vars, units)
         fig = figure(
             x_axis_label=x_label, y_axis_label=y_label, title=title,
-            width=plot_width, height=plot_height, name='radiation_daily'
-        )
+            width=plot_width, height=plot_height)
         fig = Plot.add_lines(
             fig, df, plt_vars, colors, x_label, daily_source, labels=plt_vars
         )
@@ -573,7 +568,7 @@ class Plot(object):
             title = 'Monthly Radiation Components'
             fig = figure(
                 x_axis_label=x_label, y_axis_label=y_label, title=title,
-                width=plot_width, height=plot_height, name='radiation_monthly'
+                width=plot_width, height=plot_height
             )
             fig = Plot.add_lines(
                 fig, monthly_df, plt_vars, colors, x_label, monthly_source,
@@ -592,7 +587,7 @@ class Plot(object):
         y_label = _get_units(plt_vars, units)
         fig = figure(
             x_axis_label=x_label, y_axis_label=y_label, title=title,
-            width=plot_width, height=plot_height, name='temp_daily'
+            width=plot_width, height=plot_height
         )
         fig = Plot.add_lines(fig, df, plt_vars, colors, x_label, daily_source)
[…]
