### 2021-09-19 0677cd5
further removal of name kwarg from bokeh for github test issue
