### 2021-09-20 fddd2f5
remove undefined reference related to last fix
