### 2021-09-20 da4d424
fix for new bokeh version plot text argument default
