+++ fluxdataqaqc/plot.py
@@ -199,6 +199,7 @@ class Plot(object):
             This method is also available from the :obj:`.Data` and :obj:`.QaQc`
             objects.
         """
+        name = '{}_vs_{}'.format(x,y)
         # remove pairs where one or both are nans, for LSRL and min/max-1:1
         
         if not y in source.data:
@@ -213,7 +214,7 @@ class Plot(object):
         yd = source.data[y].astype(float)
         mask = (~np.isnan(xd) & ~np.isnan(yd))
         if not mask.any():
-            print(f'WARNING: cannot plot {x} vs {y} because no paired data')
+            print(f'WARNING: cannot plot {name} because no paired data')
             return
 
         xd = xd[mask]
@@ -230,13 +231,13 @@ class Plot(object):
             fig.scatter(
                 x, y, source=source, color=color, line_width=1, fill_alpha=in_a,
                 legend_label='{lab}, slope={s:.2f}'.format(lab=label, s=m), 
-                size=size, **kwargs
+                name=name, size=size, **kwargs
             )
             fig.line(xd, m * xd, color=color)
         else:
             fig.scatter(
                 x, y, source=source, color=color, line_width=1, fill_alpha=in_a,
-                legend_label=label, size=size, **kwargs
+                legend_label=label, name=name, size=size, **kwargs
             )
         if len(fig.hover) == 0:
             if x == date_name:
@@ -266,7 +267,7 @@ class Plot(object):
             for t in new_tips:
                 if not t in fig.hover[0].tooltips:
                     fig.hover[0].tooltips.append(t)
-
+        fig.hover[0].names.append(name)
         fig.legend.location = "top_left"
         fig.legend.click_policy="hide"
         
@@ -359,7 +360,7 @@ class Plot(object):
         return ret
 
     def _plot(self, FluxObj, ncols=1, output_type='save', out_file=None, 
-            suptitle=None, plot_width=1000, plot_height=450, 
+            suptitle='', plot_width=1000, plot_height=450, 
             sizing_mode='scale_both', merge_tools=False, link_x=True, **kwargs): 
         """ 
         Private routine for aggregated validation plots that are used by
@@ -431,7 +432,7 @@ class Plot(object):
         y_label = _get_units(plt_vars, units)
         fig = figure(
             x_axis_label=x_label, y_axis_label=y_label, title=title,
-            width=plot_width, height=plot_height, 
+            width=plot_width, height=plot_height, name='energy_balance_daily'
         )
         fig = Plot.add_lines(
             fig, df, plt_vars, colors, x_label, daily_source, labels=plt_vars
@@ -448,6 +449,7 @@ class Plot(object):
             title = 'Monthly Surface Energy Balance Components'
             fig = figure(x_axis_label=x_label, y_axis_label=y_label,title=title,
                 width=plot_width, height=plot_height, 
+                name='energy_balance_monthly'
             )
             fig = Plot.add_lines(
                 fig, monthly_df, plt_vars, colors, x_label, monthly_source,
@@ -470,6 +472,7 @@ class Plot(object):
             y_label = _get_units(plt_vars, units)
             fig = figure(x_axis_label=x_label, y_axis_label=y_label, 
                 title=title, width=plot_width, height=plot_height, 
+                name='Rs_daily'
             )
             fig = Plot.add_lines(
                 fig, df, plt_vars, colors, x_label, daily_source, labels=labels
@@ -513,7 +516,7 @@ class Plot(object):
             y_label = _get_units(g_vars, units)
             fig = figure(
                 x_axis_label=x_label, y_axis_label=y_label, title=title,
-                plot_width=plot_width, plot_height=plot_height
+                plot_width=plot_width, plot_height=plot_height, name='G_daily'
             )
[…]
