### 2021-09-20 9a07813
fixes bokeh text kwarg not accepting None
