### 2021-09-20 a311b9a
shorten github action test title for display
