### 2021-09-20 #3418 @JohnVolk
I've implemented automated testing using a GitHub actions workflow [see here](https://github.com/Open-ET/flux-data-qaqc/actions/workflows/fluxdataqaqc_tests.yml) and included a badge in the README. I'll continue to add more tests to cover most of the package functionality. Thanks!
