### 2021-09-20 #3418 @pdebuyl
@JohnVolk can you describe shortly the test in the docs? There is one test file, if I understand well, that runs one typical setup and checks a few programmatic details (existence of output files) and some quantitative tests on the numerical output.
