### 2021-10-01 2750039
remove prefix from paper.bib DOIs
