+++ paper/paper.bib
@@ -32,7 +32,7 @@ volume = {156},
 pages = {65-74},
 year = {2012},
 issn = {0168-1923},
-doi = {https://doi.org/10.1016/j.agrformet.2011.12.002},
+doi = {10.1016/j.agrformet.2011.12.002},
 url = {https://www.sciencedirect.com/science/article/pii/S016819231100339X},
 author = {Ray Leuning and Eva {van Gorsel} and William J. Massman and Peter R. Isaac},
 keywords = {Energy balance closure, Eddy covariance flux measurements, Micrometeorology, Advective flux divergence},
@@ -47,7 +47,7 @@ volume = {18},
 number = {6},
 pages = {1351-1367},
 keywords = {Bowen ratio, carbon dioxide flux, energy balance closure, energy storage, latent heat flux, net radiation, scalar similarity, sensible heat flux, soil heat flux, turbulent flux},
-doi = {https://doi.org/10.1890/06-0922.1},
+doi = {10.1890/06-0922.1},
 url = {https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/06-0922.1},
 eprint = {https://esajournals.onlinelibrary.wiley.com/doi/pdf/10.1890/06-0922.1},
 abstract = {This paper gives an overview of 20 years of research on the energy balance closure problem. It will be shown that former assumptions that measuring errors or storage terms are the reason for the unclosed energy balance do not stand up because even turbulent fluxes derived from documented methods and calibrated sensors, net radiation, and ground heat fluxes cannot close the energy balance. Instead, exchange processes on larger scales of the heterogeneous landscape have a significant influence. By including these fluxes, the energy balance can be approximately closed. Therefore, the problem is a scale problem and has important consequences to the measurement and modeling of turbulent fluxes.},
@@ -73,6 +73,7 @@ eprint = {https://ascelibrary.org/doi/pdf/10.1061/9780784408056}
   number={5},
   pages={1331--1340},
   year={1988},
+  doi={10.2307/1941631},
   publisher={Wiley Online Library}
 }
 
@@ -108,7 +109,7 @@ volume = {33},
 number = {1},
 pages = {121-131},
 keywords = {weather data, humidity, agriculture, wildfire},
-doi = {https://doi.org/10.1002/joc.3413},
+doi = {10.1002/joc.3413},
 url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/joc.3413},
 eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/joc.3413},
 year = {2013}
@@ -123,7 +124,7 @@ pages = {223-243},
 year = {2002},
 note = {FLUXNET 2000 Synthesis},
 issn = {0168-1923},
-doi = {https://doi.org/10.1016/S0168-1923(02)00109-0},
+doi = {10.1016/S0168-1923(02)00109-0},
 url = {https://www.sciencedirect.com/science/article/pii/S0168192302001090},
 author = {Kell Wilson and Allen Goldstein and Eva Falge and Marc Aubinet and Dennis Baldocchi and Paul Berbigier and Christian Bernhofer and Reinhart Ceulemans and Han Dolman and Chris Field and Achim Grelle and Andreas Ibrom and B.E Law and Andy Kowalski and Tilden Meyers and John Moncrieff and Russ Monson and Walter Oechel and John Tenhunen and Riccardo Valentini and Shashi Verma},
 keywords = {Energy balance, FLUXNET, Eddy covariance technique},
@@ -149,7 +150,7 @@ volume = {171-172},
 pages = {137-152},
 year = {2013},
 issn = {0168-1923},
[…]
