### 2021-10-01 #3418 @JohnVolk
@ashwinvis I just updated the DOIs to remove those prefixes as recommended.

@pdebuyl I'll get add documentation for the testing soon. You are correct, the tests create a few different runs of the software using different example input files which are provided with the software, and checks if calculations and output files were performed/generated correctly in each case.
