### 2021-10-03 Open-ET/flux-data-qaqc#9 @ashwinvis
Basic usage tutorial errors
```py
>>> d.plot(output_type='notebook', plot_width=700)
NameError                                 Traceback (most recent call last)
/tmp/ipykernel_793706/3470416942.py in <module>
----> 1 d.plot(output_type='notebook', plot_width=700)

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/fluxdataqaqc/data.py in plot(self, ncols, output_type, out_file, suptitle, plot_width, plot_height, sizing_mode, merge_tools, link_x, **kwargs)
    537
    538         # create aggregrated plot structure from fluxdataqaqc.Plot._plot()
--> 539         ret = self._plot(
    540             self, ncols=ncols, output_type=output_type, out_file=out_file,
    541             suptitle=suptitle, plot_width=plot_width, plot_height=plot_height,

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/fluxdataqaqc/plot.py in _plot(self, FluxObj, ncols, output_type, out_file, suptitle, plot_width, plot_height, sizing_mode, merge_tools, link_x, **kwargs)
   1172             if fig is not None:
   1173                 daily_line.append(fig)
-> 1174             theta_vars = [
   1175                 v for v in variables if theta_re.match(v) and v in\
   1176                     monthly_df.columns

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/fluxdataqaqc/plot.py in <listcomp>(.0)
   1174             theta_vars = [
   1175                 v for v in variables if theta_re.match(v) and v in\
-> 1176                     monthly_df.columns
   1177             ]
   1178             if fig is not None and monthly and len(theta_vars) > 0:

NameError: free variable 'monthly_df' referenced before assignment in enclosing scope
```

```py
>>> # creating a QaQc instance will automatically convert to daily
>>> d = Data('US-Tw3_config.ini')
>>> q = QaQc(d)
IndexError                                Traceback (most recent call last)
/tmp/ipykernel_793706/3547759801.py in <module>
      1 # creating a QaQc instance will automatically convert to daily
      2 d = Data('US-Tw3_config.ini')
----> 3 q = QaQc(d)

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/fluxdataqaqc/qaqc.py in __init__(self, data, drop_gaps, daily_frac, max_interp_hours, max_interp_hours_night)
    288
    289             # data will be loaded if it has not yet via Data.df
--> 290             self.temporal_freq = self._check_daily_freq(
    291                 drop_gaps, daily_frac, max_interp_hours, max_interp_hours_night
    292             )

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/fluxdataqaqc/qaqc.py in _check_daily_freq(self, drop_gaps, daily_frac, max_interp_hours, max_interp_hours_night)
    785             self.n_samples_per_day = 1
    786
--> 787         self._df = df.rename(self.variables)
    788         return freq
    789

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/util/_decorators.py in wrapper(*args, **kwargs)
    322         @wraps(func)
    323         def wrapper(*args, **kwargs) -> Callable[..., Any]:
--> 324             return func(*args, **kwargs)
    325
    326         kind = inspect.Parameter.POSITIONAL_OR_KEYWORD

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/frame.py in rename(self, mapper, index, columns, axis, copy, inplace, level, errors)
   5032         4  3  6
   5033         """
-> 5034         return super().rename(
   5035             mapper=mapper,
   5036             index=index,

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/generic.py in rename(self, mapper, index, columns, axis, copy, inplace, level, errors)
   1145             # GH 13473
   1146             if not callable(replacements):
-> 1147                 indexer = ax.get_indexer_for(replacements)
   1148                 if errors == "raise" and len(indexer[indexer == -1]):
   1149                     missing_labels = [

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/indexes/base.py in get_indexer_for(self, target, **kwargs)
   5274         """
   5275         if self._index_as_unique:
-> 5276             return self.get_indexer(target, **kwargs)
   5277         indexer, _ = self.get_indexer_non_unique(target)
   5278         return indexer

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/indexes/base.py in get_indexer(self, target, method, limit, tolerance)
   3435         # returned ndarray is np.intp
   3436         method = missing.clean_reindex_fill_method(method)
-> 3437         target = self._maybe_cast_listlike_indexer(target)
   3438
   3439         self._check_indexing_method(method, limit, tolerance)

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/indexes/datetimelike.py in _maybe_cast_listlike_indexer(self, keyarr)
    599     def _maybe_cast_listlike_indexer(self, keyarr):
    600         try:
--> 601             res = self._data._validate_listlike(keyarr, allow_object=True)
    602         except (ValueError, TypeError):
    603             res = com.asarray_tuplesafe(keyarr)

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/arrays/datetimelike.py in _validate_listlike(self, value, allow_object)
    701         # Do type inference if necessary up front
    702         # e.g. we passed PeriodIndex.values and got an ndarray of Periods
--> 703         value = pd_array(value)
    704         value = extract_array(value, extract_numpy=True)
    705

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/construction.py in array(data, dtype, copy)
    344         elif inferred_dtype == "string":
    345             # StringArray/ArrowStringArray depending on pd.options.mode.string_storage
--> 346             return StringDtype().construct_array_type()._from_sequence(data, copy=copy)
    347
    348         elif inferred_dtype == "integer":

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/core/arrays/string_.py in _from_sequence(cls, scalars, dtype, copy)
    345         else:
    346             # convert non-na-likes to str, and nan-likes to StringDtype.na_value
--> 347             result = lib.ensure_string_array(
    348                 scalars, na_value=StringDtype.na_value, copy=copy
    349             )

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/_libs/lib.pyx in pandas._libs.lib.ensure_string_array()

~/src/sandbox/flux-data-qaqc/venv/lib/python3.8/site-packages/pandas/_libs/lib.pyx in pandas._libs.lib.ensure_string_array()

IndexError: too many indices for array
```
