### 2021-10-03 Open-ET/flux-data-qaqc#10 @ashwinvis
Low code coverage
If I run

```
❯ pytest --cov=fluxdataqaqc tests/
==================================== test session starts ====================================
rootdir: /home/avmo/src/sandbox/flux-data-qaqc, configfile: pytest.ini
plugins: cov-2.12.1, nbval-0.9.6
collected 11 items

tests/test_fluxdataqaqc.py ...........                                                [100%]

----------- coverage: platform linux, python 3.8.9-final-0 -----------
Name                       Stmts   Miss  Cover
----------------------------------------------
fluxdataqaqc/__init__.py       8      0   100%
fluxdataqaqc/data.py         488    114    77%
fluxdataqaqc/plot.py         552    290    47%
fluxdataqaqc/qaqc.py         663    619     7%
fluxdataqaqc/util.py          90     60    33%
----------------------------------------------
TOTAL                       1801   1083    40%

==================================== 11 passed in 21.02s ====================================
```

I notice that the tests mostly target only the `data` module. If the tutorials cover the rest then it should be good enough and one could check the coverage as follows.
```
❯ cd examples/Basic_usage
❯ pytest --cov=fluxdataqaqc --nbval Tutorial.ipynb ../../tests/
```
However, in order to do that, #9 should be fixed first.
