### 2021-10-03 #3418 @ashwinvis
# Minor revisions needed

<!DOCTYPE html>
<html><head>
		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
		<link rel="stylesheet" type="text/css" href="data:text/css;base64,Ym9keSB7CgliYWNrZ3JvdW5kOiB3aGl0ZTsKfQoKYSB7Cgl0ZXh0LWRlY29yYXRpb246IHVuZGVybGluZTsKfQoKYm9keSB7CglwYWRkaW5nOiAwOwp9Cgp1bC5yZXBvcnQgbGkuaXRlbSB7Cglib3JkZXItdG9wOiA0cHggc29saWQgIzU1NTsKCXBhZGRpbmctdG9wOiAxZW07CglwYWRkaW5nLWxlZnQ6IDFlbTsKCXBhZGRpbmctcmlnaHQ6IDFlbTsKCW1hcmdpbi1ib3R0b206IDJlbTsKfQoKaDEsIGgyLCBoMywgaDQsIGg1LCBoNiB7Cglmb250LXdlaWdodDogbm9ybWFsOwp9CgpoMiB7CgltYXJnaW46IDAgMCAuNWVtOwp9CgpoMi5wYXJlbnRJdGVtIHsKCWZvbnQtd2VpZ2h0OiBib2xkOwoJZm9udC1zaXplOiAxZW07CglwYWRkaW5nOiAwIDAgLjVlbTsKCWJvcmRlci1ib3R0b206IDFweCBzb2xpZCAjY2NjOwp9CgovKiBJZiBjb21iaW5pbmcgY2hpbGRyZW4sIGRpc3BsYXkgcGFyZW50IHNsaWdodGx5IGxhcmdlciAqLwp1bC5yZXBvcnQuY29tYmluZUNoaWxkSXRlbXMgaDIucGFyZW50SXRlbSB7Cglmb250LXNpemU6IDEuMWVtOwoJcGFkZGluZy1ib3R0b206IC43NWVtOwoJbWFyZ2luLWJvdHRvbTogLjRlbTsKfQoKaDIucGFyZW50SXRlbSAudGl0bGUgewoJZm9udC13ZWlnaHQ6IG5vcm1hbDsKfQoKaDMgewoJbWFyZ2luLWJvdHRvbTogLjZlbTsKCWZvbnQtd2VpZ2h0OiBib2xkICFpbXBvcnRhbnQ7Cglmb250LXNpemU6IDFlbTsKCWRpc3BsYXk6IGJsb2NrOwp9CgovKiBNZXRhZGF0YSB0YWJsZSAqLwp0aCB7Cgl2ZXJ0aWNhbC1hbGlnbjogdG9wOwoJdGV4dC1hbGlnbjogcmlnaHQ7Cgl3aWR0aDogMTUlOwoJd2hpdGUtc3BhY2U6IG5vd3JhcDsKfQoKdGQgewoJcGFkZGluZy1sZWZ0OiAuNWVtOwp9CgoKdWwucmVwb3J0LCB1bC5ub3RlcywgdWwudGFncyB7CglsaXN0LXN0eWxlOiBub25lOwoJbWFyZ2luLWxlZnQ6IDA7CglwYWRkaW5nLWxlZnQ6IDA7Cn0KCi8qIFRhZ3MgKi8KaDMudGFncyB7Cglmb250LXNpemU6IDEuMWVtOwp9Cgp1bC50YWdzIHsKCWxpbmUtaGVpZ2h0OiAxLjc1ZW07CglsaXN0LXN0eWxlOiBub25lOwp9Cgp1bC50YWdzIGxpIHsKCWRpc3BsYXk6IGlubGluZTsKfQoKdWwudGFncyBsaTpub3QoOmxhc3QtY2hpbGQpOmFmdGVyIHsKCWNvbnRlbnQ6ICcsICc7Cn0KCgovKiBDaGlsZCBub3RlcyAqLwpoMy5ub3RlcyB7Cglmb250LXNpemU6IDEuMWVtOwp9Cgp1bC5ub3RlcyB7CgltYXJnaW4tYm90dG9tOiAxLjJlbTsKfQoKdWwubm90ZXMgPiBsaTpmaXJzdC1jaGlsZCBwIHsKCW1hcmdpbi10b3A6IDA7Cn0KCnVsLm5vdGVzID4gbGkgewoJcGFkZGluZzogLjdlbSAwOwp9Cgp1bC5ub3RlcyA+IGxpOm5vdCg6bGFzdC1jaGlsZCkgewoJYm9yZGVyLWJvdHRvbTogMXB4ICNjY2Mgc29saWQ7Cn0KCgp1bC5ub3RlcyA+IGxpIHA6Zmlyc3QtY2hpbGQgewoJbWFyZ2luLXRvcDogMDsKfQoKdWwubm90ZXMgPiBsaSBwOmxhc3QtY2hpbGQgewoJbWFyZ2luLWJvdHRvbTogMDsKfQoKLyogQWRkIHF1b3RhdGlvbiBtYXJrcyBhcm91bmQgYmxvY2txdW90ZSAqLwp1bC5ub3RlcyA+IGxpIGJsb2NrcXVvdGUgcDpub3QoOmVtcHR5KTpiZWZvcmUsCmxpLm5vdGUgYmxvY2txdW90ZSBwOm5vdCg6ZW1wdHkpOmJlZm9yZSB7Cgljb250ZW50OiAn4oCcJzsKfQoKdWwubm90ZXMgPiBsaSBibG9ja3F1b3RlIHA6bm90KDplbXB0eSk6bGFzdC1jaGlsZDphZnRlciwKbGkubm90ZSBibG9ja3F1b3RlIHA6bm90KDplbXB0eSk6bGFzdC1jaGlsZDphZnRlciB7Cgljb250ZW50OiAn4oCdJzsKfQoKLyogUHJlc2VydmUgd2hpdGVzcGFjZSBvbiBwbGFpbnRleHQgbm90ZXMgKi8KdWwubm90ZXMgbGkgcC5wbGFpbnRleHQsIGxpLm5vdGUgcC5wbGFpbnRleHQsIGRpdi5ub3RlIHAucGxhaW50ZXh0IHsKCXdoaXRlLXNwYWNlOiBwcmUtd3JhcDsKfQoKLyogRGlzcGxheSB0YWdzIHdpdGhpbiBjaGlsZCBub3RlcyBpbmxpbmUgKi8KdWwubm90ZXMgaDMudGFncyB7CglkaXNwbGF5OiBpbmxpbmU7Cglmb250LXNpemU6IDFlbTsKfQoKdWwubm90ZXMgaDMudGFnczphZnRlciB7Cgljb250ZW50OiAnICc7Cn0KCnVsLm5vdGVzIHVsLnRhZ3MgewoJZGlzcGxheTogaW5saW5lOwp9Cgp1bC5ub3RlcyB1bC50YWdzIGxpOm5vdCg6bGFzdC1jaGlsZCk6YWZ0ZXIgewoJY29udGVudDogJywgJzsKfQoKCi8qIENoaWxkIGF0dGFjaG1lbnRzICovCmgzLmF0dGFjaG1lbnRzIHsKCWZvbnQtc2l6ZTogMS4xZW07Cn0KCnVsLmF0dGFjaG1lbnRzIGxpIHsKCXBhZGRpbmctdG9wOiAuNWVtOwp9Cgp1bC5hdHRhY2htZW50cyBkaXYubm90ZSB7CgltYXJnaW4tbGVmdDogMmVtOwp9Cgp1bC5hdHRhY2htZW50cyBkaXYubm90ZSBwOmZpcnN0LWNoaWxkIHsKCW1hcmdpbi10b3A6IC43NWVtOwp9Cg==">
		<link rel="stylesheet" type="text/css" media="screen,projection" href="data:text/css;base64,LyogR2VuZXJpYyBzdHlsZXMgKi8KYm9keSB7Cglmb250OiA2Mi41JSBHZW9yZ2lhLCBUaW1lcywgc2VyaWY7Cgl3aWR0aDogNzgwcHg7CgltYXJnaW46IDAgYXV0bzsKfQoKaDIgewoJZm9udC1zaXplOiAxLjVlbTsKCWxpbmUtaGVpZ2h0OiAxLjVlbTsKCWZvbnQtZmFtaWx5OiBHZW9yZ2lhLCBUaW1lcywgc2VyaWY7Cn0KCnAgewoJbGluZS1oZWlnaHQ6IDEuNWVtOwp9CgphOmxpbmssIGE6dmlzaXRlZCB7Cgljb2xvcjogIzkwMDsKfQoKYTpob3ZlciwgYTphY3RpdmUgewoJY29sb3I6ICM3Nzc7Cn0KCgp1bC5yZXBvcnQgewoJZm9udC1zaXplOiAxLjRlbTsKCXdpZHRoOiA2ODBweDsKCW1hcmdpbjogMCBhdXRvOwoJcGFkZGluZzogMjBweCAyMHB4Owp9CgovKiBNZXRhZGF0YSB0YWJsZSAqLwp0YWJsZSB7Cglib3JkZXI6IDFweCAjY2NjIHNvbGlkOwoJb3ZlcmZsb3c6IGF1dG87Cgl3aWR0aDogMTAwJTsKCW1hcmdpbjogLjFlbSBhdXRvIC43NWVtOwoJcGFkZGluZzogMC41ZW07Cn0K">
		<link rel="stylesheet" type="text/css" media="print" href="data:text/css;base64,Ym9keSB7Cglmb250OiAxMnB0ICJUaW1lcyBOZXcgUm9tYW4iLCBUaW1lcywgR2VvcmdpYSwgc2VyaWY7CgltYXJnaW46IDA7Cgl3aWR0aDogYXV0bzsKCWNvbG9yOiBibGFjazsKfQoKLyogUGFnZSBCcmVha3MgKHBhZ2UtYnJlYWstaW5zaWRlIG9ubHkgcmVjb2duaXplZCBieSBPcGVyYSkgKi8KaDEsIGgyLCBoMywgaDQsIGg1LCBoNiB7CglwYWdlLWJyZWFrLWFmdGVyOiBhdm9pZDsKCXBhZ2UtYnJlYWstaW5zaWRlOiBhdm9pZDsKfQoKdWwsIG9sLCBkbCB7CglwYWdlLWJyZWFrLWluc2lkZTogYXZvaWQ7Cgljb2xvci1hZGp1c3Q6IGV4YWN0Owp9CgpoMiB7Cglmb250LXNpemU6IDEuM2VtOwoJbGluZS1oZWlnaHQ6IDEuM2VtOwp9CgphIHsKCWNvbG9yOiAjMDAwOwoJdGV4dC1kZWNvcmF0aW9uOiBub25lOwp9Cg==">
	</head>
	<body>
		<ul class="report combineChildItems">
			<li id="item_R4U8IA4C" class="item journalArticle">
				<ul class="notes">
					<li id="item_UWIDWYQN">
<div><div data-citation-items="%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22itemData%22%3A%7B%22id%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%2C%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FVJMBWKJK%22%5D%2C%22itemData%22%3A%7B%22id%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FVJMBWKJK%22%2C%22type%22%3A%22webpage%22%2C%22title%22%3A%22The%20FLUXNET2015%20dataset%20and%20the%20ONEFlux%20processing%20pipeline%20for%20eddy%20covariance%20data%20%7C%20Scientific%20Data%20%7C%20https%3A%2F%2Fwww.nature.com%2F%22%2C%22URL%22%3A%22https%3A%2F%2Fwww.nature.com%2Farticles%2Fs41597-020-0534-3%22%2C%22accessed%22%3A%7B%22date-parts%22%3A%5B%5B%222021%22%2C10%2C3%5D%5D%7D%7D%7D%2C%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FNFEAFTVD%22%5D%2C%22itemData%22%3A%7B%22id%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FNFEAFTVD%22%2C%22type%22%3A%22webpage%22%2C%22title%22%3A%22Friction%20velocity%20-%20Glossary%20of%20Meteorology%20%7C%20https%3A%2F%2Fglossary.ametsoc.org%2F%22%2C%22URL%22%3A%22https%3A%2F%2Fglossary.ametsoc.org%2Fwiki%2FFriction_velocity%22%2C%22accessed%22%3A%7B%22date-parts%22%3A%5B%5B%222021%22%2C10%2C3%5D%5D%7D%7D%7D%5D" data-schema-version="4"><h1>Software paper</h1>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%221%22%2C%22position%22%3A%7B%22pageIndex%22%3A0%2C%22rects%22%3A%5B%5B166.864%2C294.841%2C552.748%2C305.083%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%221%22%7D%7D">“The flux-data-qaqc open-source Python package provides a standard and reproducible”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%221%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 1)</span> <br>Which
 standards are we referring to? Does one exist? If yes, please cite
them. If no, avoid the adjective "standard". It gives the impression
that all the methods follows some standard.<br><br>Also, reproduciblity
is a high bar to claim, both as a method and as a software. It appears
to me that there are inherent uncertainties in the method (cf. <strong>Introduction)</strong>;
 can we call the method reproducible? If it is for the software
toolchain, how do you ensure reproduciblity? It is good to substantiate
this claim.</p>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%221%22%2C%22position%22%3A%7B%22pageIndex%22%3A0%2C%22rects%22%3A%5B%5B166.864%2C211.433%2C477.647%2C221.396%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%221%22%7D%7D">“provides a reproducible framework for common post-processing methods”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%221%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 1)</span> Again, "reproducible"</p>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%221%22%2C%22position%22%3A%7B%22pageIndex%22%3A0%2C%22rects%22%3A%5B%5B362.818%2C175.568%2C552.755%2C185.531%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%221%22%7D%7D">“Post-processing capabilities and tools provided”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%221%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 1)</span>
 I wonder if this bullet list should be moved to the next section
"Design and features"? Are all these features relevant to the "Statement
 of need" or nice things this package provides?</p>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%222%22%2C%22position%22%3A%7B%22pageIndex%22%3A1%2C%22rects%22%3A%5B%5B191.771%2C623.185%2C552.746%2C633.148%5D%2C%5B151.612%2C624.152%2C156.902%2C629.133%5D%2C%5B191.771%2C611.23%2C530.759%2C621.193%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%222%22%7D%7D">“other
 atmospheric calculations, including the hourly and daily forms of the
American 48 Society of Civil Engineers standardized reference ET
equation (Allen et al., 2005)”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%222%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 2)</span>
 Is this the "standard" referred to in Page 1. Do you have another
reference which is openly accessible? My university does not have access
 to this, and I suppose many won't be able to read this either.</p>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%222%22%2C%22position%22%3A%7B%22pageIndex%22%3A1%2C%22rects%22%3A%5B%5B512.447%2C395.591%2C539.854%2C405.554%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%222%22%7D%7D">“QAQC”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%222%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 2)</span> Definition for QAQC missing. Isn't it styled as QA/QC as in one of the papers you cited <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FVJMBWKJK%22%5D%2C%22itemData%22%3A%7B%22id%22%3A6529%2C%22type%22%3A%22webpage%22%2C%22title%22%3A%22The%20FLUXNET2015%20dataset%20and%20the%20ONEFlux%20processing%20pipeline%20for%20eddy%20covariance%20data%20%7C%20Scientific%20Data%20%7C%20https%3A%2F%2Fwww.nature.com%2F%22%2C%22URL%22%3A%22https%3A%2F%2Fwww.nature.com%2Farticles%2Fs41597-020-0534-3%22%2C%22accessed%22%3A%7B%22date-parts%22%3A%5B%5B%222021%22%2C10%2C3%5D%5D%7D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(“The
 FLUXNET2015 dataset and the ONEFlux processing pipeline for eddy
covariance data | Scientific Data | https://www.nature.com/”)</span>?</p>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%223%22%2C%22position%22%3A%7B%22pageIndex%22%3A2%2C%22rects%22%3A%5B%5B464.24%2C463.907%2C552.756%2C472.873%5D%2C%5B166.864%2C452.948%2C307.045%2C461.914%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%223%22%7D%7D">“arrows signify the data flow pipeline and class inheritance,”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%223%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 3)</span> Use different colour or style for arrows and distinguish them in the caption / legend.</p>
<p><span class="highlight" data-annotation="%7B%22attachmentURI%22%3A%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FXENNFY6M%22%2C%22color%22%3A%22%23ffd400%22%2C%22pageLabel%22%3A%223%22%2C%22position%22%3A%7B%22pageIndex%22%3A2%2C%22rects%22%3A%5B%5B276.343%2C396.664%2C552.745%2C406.627%5D%2C%5B151.612%2C397.631%2C156.902%2C402.612%5D%2C%5B166.864%2C384.43%2C418.918%2C394.672%5D%5D%7D%2C%22citationItem%22%3A%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%223%22%7D%7D">“configuration
 file that can be written by the user or using a tool 66 provided with
the package, and full example files are provided”</span> <span class="citation" data-citation="%7B%22citationItems%22%3A%5B%7B%22uris%22%3A%5B%22http%3A%2F%2Fzotero.org%2Fusers%2F3909932%2Fitems%2FR4U8IA4C%22%5D%2C%22locator%22%3A%223%22%2C%22itemData%22%3A%7B%22id%22%3A6468%2C%22type%22%3A%22article-journal%22%2C%22language%22%3A%22en%22%2C%22page%22%3A%225%22%2C%22source%22%3A%22Zotero%22%2C%22title%22%3A%22flux-data-qaqc%3A%20A%20Python%20Package%20for%20Energy%20Balance%20Closure%20and%20Post-Processing%20of%20Eddy%20Flux%20Data%22%2C%22author%22%3A%5B%7B%22family%22%3A%22Volk%22%2C%22given%22%3A%22John%22%7D%2C%7B%22family%22%3A%22Huntington%22%2C%22given%22%3A%22Justin%22%7D%2C%7B%22family%22%3A%22Allen%22%2C%22given%22%3A%22Richard%22%7D%2C%7B%22family%22%3A%22Melton%22%2C%22given%22%3A%22Forrest%22%7D%2C%7B%22family%22%3A%22Anderson%22%2C%22given%22%3A%22Martha%22%7D%2C%7B%22family%22%3A%22Kilic%22%2C%22given%22%3A%22Ayse%22%7D%5D%7D%7D%5D%2C%22properties%22%3A%7B%7D%7D">(Volk et al., p. 3)</span> Where is it provided? Link to documentation or source code.</p>
[…]
