### 2021-10-04 Open-ET/flux-data-qaqc#9 @JohnVolk
The error that you received should of been fixed in commit c9db99cc0db45ee7015c3420c5a31314c60fd261, I just added a new release to PyPI in case you were installing from there.
