### 2021-10-04 2cc1824
addresses bug found in #9
