### 2021-10-04 #3418 @pdebuyl
Thank you @ashwinvis for the review.

I see that @JohnVolk started to reply in the issues, I'll check them but it is easier for me if the resolutions of the issues are mentioned here.
