### 2021-10-04 Open-ET/flux-data-qaqc#9 @ashwinvis
I am using pandas v1.3.2, if it helps.
