+++ paper/paper.bib
@@ -13,6 +13,7 @@ DOI = {10.5194/bg-15-5015-2018}
 @article{kustas2018,
  ISSN = {00030007, 15200477},
  URL = {https://www.jstor.org/stable/26639471},
+ DOI = {10.1175/BAMS-D-16-0244.1},
  abstract = {Particularly in light of California’s recent multiyear drought, there is a critical need for accurate and timely evapotranspiration (ET) and crop stress information to ensure long-term sustainability of high-value crops. Providing this information requires the development of tools applicable across the continuum from subfield scales to improve water management within individual fields up to watershed and regional scales to assess water resources at county and state levels. High-value perennial crops (vineyards and orchards) are major water users, and growers will need better tools to improve water-use efficiency to remain economically viable and sustainable during periods of prolonged drought. To develop these tools, government, university, and industry partners are evaluating a multiscale remote sensing–based modeling system for application over vineyards. During the 2013–17 growing seasons, the Grape Remote Sensing Atmospheric Profile and Evapotranspiration eXperiment (GRAPEX) project has collected micrometeorological and biophysical data within adjacent pinot noir vineyards in the Central Valley of California. Additionally, each year ground, airborne, and satellite remote sensing data were collected during intensive observation periods (IOPs) representing different vine phenological stages. An overview of the measurements and some initial results regarding the impact of vine canopy architecture on modeling ET and plant stress are presented here. Refinements to the ET modeling system based on GRAPEX are being implemented initially at the field scale for validation and then will be integrated into the regional modeling toolkit for large area assessment.},
  author = {William P. Kustas and Martha C. Anderson and Joseph G. Alfieri and Kyle Knipper and Alfonso Torres-Rua and Christopher K. Parry and Hector Nieto and Nurit Agam and William A. White and Feng Gao and Lynn McKee and John H. Prueger and Lawrence E. Hipps and Sebastian Los and Maria Mar Alsina and Luis Sanchez and Brent Sams and Nick Dokoozlian and Mac McKee and Scott Jones and Yun Yang and Tiffany G. Wilson and Fangni Lei and Andrew McElrone and Josh L. Heitman and Adam M. Howard and Kirk Post and Forrest Melton and Christopher Hain},
  journal = {Bulletin of the American Meteorological Society},
