+++ paper/paper.md
@@ -39,11 +39,23 @@ bibliography: paper.bib
 
 # Introduction
 
-Eddy Covariance (EC) systems provide continuous direct measurements of turbulent and radiative heat and energy fluxes between the atmosphere and land surface [@baldocchi1988]. Evapotranspiration (ET) estimates from EC systems are highly valuable for many scientific disciplines, water policy decisions, and operational applications such as irrigation management. Although the EC technique is considered accurate under certain atmospheric and spatial conditions, inherent theoretical and practical limitations of the technique are magnified by data uncertainty from post-processing methods. Error can result from the well-documented energy balance closure problem resulting from inherent limitations in the technique but also from data processing errors, instrumentation errors, and other sources [@foken2008; @stoy2013]. Several approaches for post-processing and correcting EC-based  ET estimates exist, notably those implemented by major flux measurement networks and datasets like FLUXNET2015 [@pastorello2020]. However, post-processing routines often differ between individual site teams, data networks, and end users, and in some cases post-processing decisions are difficult to reproduce. 
+Eddy Covariance (EC) systems provide continuous direct measurements of turbulent and radiative heat and energy fluxes between the atmosphere and land surface [@baldocchi1988]. Evapotranspiration (ET) estimates from EC systems are highly valuable for many scientific disciplines, water policy decisions, and operational applications such as irrigation management. Although the EC technique is considered accurate under certain atmospheric and spatial conditions, inherent theoretical and practical limitations of the technique are magnified by data uncertainty from post-processing methods. Error can result from the well-documented energy balance closure problem resulting from inherent limitations in the technique but also from data processing errors, instrumentation errors, and other sources [@foken2008; @stoy2013]. Several approaches for post-processing and correcting EC-based ET estimates exist, notably those implemented by major flux measurement networks and datasets like FLUXNET2015 [@pastorello2020]. However, post-processing routines often differ between individual site teams, data networks, and end users, and in some cases post-processing decisions are difficult to reproduce. 
 
 # Statement of need
 
-The ``flux-data-qaqc`` open-source Python package provides a standard and reproducible framework for daily energy balance closure corrections and common post-processing steps applied to EC and meteorological data. Ultimately, obtaining final ET estimates from EC data involves multiple human-driven processing decisions and this presents an uncertainty issue when comparing data that are processed using different methods. ``flux-data-qaqc`` does not address issues regarding the deployment configuration  of the EC system or the initial processing of high frequency EC data, for which several software exist. Instead, this software provides a reproducible framework for common post-processing methods (e.g., gap-filling, temporal aggregation) and energy balance closure assessment [@wilson2002]. It is focused towards processing daily and monthly ET from input data at  half-hourly or hourly frequency, such as from the AmeriFlux dataset. Post-processing capabilities and tools provided by ``flux-data-qaqc`` include:
[…]
