### 2021-10-04 #3418 @JohnVolk
Thank you @ashwinvis for the useful comments and suggestions. I've addressed most of the points related to the software paper in the last commit except at this point I cannot find another option for the Allen et al., 2005 reference. Still working on the documentation and automated test coverage. Regarding the Allen et al., 2005 report, it is specifically written around the equation and is the most appropriate citation. However I am reaching out the author to see if there are any other options, he is also a coauthor on this software paper.
