### 2021-10-04 #3418 @JohnVolk
@ashwinvis, just FYI, I did a google search for the report "The ASCE Standardized Reference Evapotranspiration Equation" and the second link that comes up for me is a PDF version of [the report](https://www.mesonet.org/images/site/ASCE_Evapotranspiration_Formula.pdf).
