### 2021-10-04 #3418 @JohnVolk
Also, just in case you are interested, I just heard back from Richard Allen, he mentioned that that report is fine to use as well as [this one](https://www.uidaho.edu/-/media/UIdaho-Responsive/Files/cals/centers/Kimberly/water-resources/ASCE-Standardized-Reference-Evapotranspiration-Equation-draft-report-without-appendices.pdf?la=en&hash=F4E3EB1290A7971E172F23A2A3B660B237D727D0) on his website. They are both final drafts of the main report, the appendices are also linked on the website.  However, when I run the JOSS preprint with a URL and a DOI included in a citation, the reference style it applies does not show the URL in the bibliography, so I suppose we can drop the DOI for Allen at al. 2005 in favor of a link to the PDF in the bibliography.
