### 2021-10-04 #3418 @pdebuyl
Hi @JohnVolk indeed, DOIs take precedence. Could you try a bibtex "note" field along

> also available at URL

?

This would let readers find a freely accessible version.
