### 2021-10-05 4bbfe6a
Update docs for tests usage and grammar checks
