+++ paper/paper.bib
@@ -60,10 +60,8 @@ author = {Richard G. Allen  and Ivan A. Walter  and Ronald L. Elliott  and Terry
 title = {The ASCE Standardized Reference Evapotranspiration Equation},
 publisher = {American Society of Civil Engineers},
 year = {2005},
-doi = {10.1061/9780784408056},
-edition   = {},
-URL = {https://ascelibrary.org/doi/abs/10.1061/9780784408056},
-eprint = {https://ascelibrary.org/doi/pdf/10.1061/9780784408056}
+DOI = {10.1061/9780784408056},
+note = {also available at https://www.mesonet.org/images/site/ASCE_Evapotranspiration_Formula.pdf},
 }
 
 @article{baldocchi1988,
