### 2021-10-05 8de5ec4
remove gridMET paths from example config files
