+++ examples/Basic_usage/USGS_config.ini
@@ -5,7 +5,6 @@ station_latitude = 39.762511
 station_longitude = -117.960100
 station_elevation = 1046
 missing_data_value = -9999
-gridmet_file_path = /home/john/flux-data-qaqc/examples/Basic_usage/gridMET_data/DVD_10_39.7750N_-117.9750W.csv
 
 [DATA]
 datestring_col = Timestamp
