+++ examples/Basic_usage/fluxnet_config.ini
@@ -6,7 +6,6 @@ station_longitude = -99.42
 station_elevation = 611
 anemometer_height = 3
 missing_data_value = -9999
-gridmet_file_path = /home/john/flux-data-qaqc/examples/Basic_usage/gridMET_data/US-AR1_36.4417N_-99.4333W.csv
 
 [DATA]
 datestring_col = TIMESTAMP
