### 2021-10-05 Open-ET/flux-data-qaqc#9 @JohnVolk
That first warning is because I must have left the path to the gridMET file on my personal machine in the config.ini file but yes since the field was filled in with a path specific to my computer, that WARNING was expected.

The second error you found is due to a change in the pandas merge function with the newer version you are using, the next commit should fix it.
