### 2021-10-05 Open-ET/flux-data-qaqc#9 @ashwinvis
The tutorial runs fine now. We can mark this as solved. Thanks @JohnVolk

c.f. https://github.com/openjournals/joss-reviews/issues/3418
