### 2021-10-05 Open-ET/flux-data-qaqc#10 @ashwinvis
The code coverage is up to 72% if we include the tutorial. Of course this is not an automated unit test and there is room for improvement. For the purpose of this review (https://github.com/openjournals/joss-reviews/issues/3418) this is satisfactory.

You could pursue this further in the future @JohnVolk.
