### 2021-10-05 Open-ET/flux-data-qaqc#10 @ashwinvis
Note: Bokeh causes outputs to change, so I used the `--nbval-lax` instead.

```
pytest --cov=fluxdataqaqc --cov-report=html --nbval-lax -v Tutorial.ipynb ../../tests/
```
