### 2021-10-05 #3418 @ashwinvis
@JohnVolk I noticed that some of the key methods utilize a package called `refet` (https://github.com/WSWUP/RefET). I think this should be acknowledged. Also note somewhere in the software paper what quantities are calculated using `refet`.
