### 2021-10-05 Open-ET/flux-data-qaqc#10 @JohnVolk
Thanks for the info @ashwinvis I wasn't aware of this plugin and the options, it is really useful. I was also having issues with the --cov  option on the notebooks. Yes, I plan on flushing out the automated tests.
