### 2021-10-05 Open-ET/flux-data-qaqc#10 @ashwinvis
I used the plugins `pytest-cov` (coverage) and `nbval` (notebook testing), as you might have figured out already.
