+++ paper/paper.bib
@@ -61,7 +61,7 @@ title = {The ASCE Standardized Reference Evapotranspiration Equation},
 publisher = {American Society of Civil Engineers},
 year = {2005},
 DOI = {10.1061/9780784408056},
-note = {also available at https://www.mesonet.org/images/site/ASCE_Evapotranspiration_Formula.pdf},
+note = {also available at \url{https://www.mesonet.org/images/site/ASCE_Evapotranspiration_Formula.pdf}},
 }
 
 @article{baldocchi1988,
