+++ docs/source/contributors.rst
@@ -12,7 +12,7 @@ team members,
 * Forrest Melton
 * Dr. Gabriel Senay
 * Dr. Joshua Fisher
-* Dr. John Volk (developer, <john.volk@dri.edu>)
+* Dr. John Volk 
 * Dr. Justin Huntington
 * Dr. Martha Anderson
 * Dr. Richard G. Allen
@@ -25,6 +25,8 @@ The Energy Balance Ratio correction method was modified from the `FLUXNET2015
 sensible heat.  Additionally, FLUXNET and `AmeriFlux
 <https://ameriflux.lbl.gov/>`__ eddy covariance data was used for code
 development and we thank them for mantaining and providing eddy flux data.
+
+We would also like to acknowledge the `RefET <https://github.com/WSWUP/RefET>`__ python package and its developer Charles Morton. ``RefET`` methods are used in ``flux-data-qaqc`` for calculating the American Soceity of Civil Engineers Standardized reference ET using the hourly and daily formulations, as well as the calculation for potential solar radiation.
 
 Contributing
 ^^^^^^^^^^^^
