### 2021-10-05 #3418 @pdebuyl
@JohnVolk I'll check this citation issue. I tested it locally and it seems that we should edit our style file, which requires discussion in the JOSS team. I'll let you know when I have more info.
