### 2021-10-06 #3418 @JohnVolk
> @JohnVolk I noticed that some of the key methods utilize a package called `refet` (https://github.com/WSWUP/RefET). I think this should be acknowledged. Also note somewhere in the software paper what quantities are calculated using `refet`.

This has been addressed.
