### 2021-10-06 a0f2940
update paper statement of need, make target audience clear
