+++ paper/paper.md
@@ -43,7 +43,13 @@ Eddy Covariance (EC) systems provide continuous direct measurements of turbulent
 
 # Statement of need
 
-The ``flux-data-qaqc`` open-source Python package provides a framework for daily energy balance closure corrections and common post-processing steps applied to EC and meteorological data. Ultimately, obtaining final ET estimates from EC data involves multiple human-driven processing decisions and this presents an uncertainty issue when comparing data that are processed using different methods. ``flux-data-qaqc`` does not address issues regarding the deployment configuration  of the EC system or the initial processing of high frequency EC data, for which several software exist. Instead, this software provides a framework and provenance for EC data post-processing (e.g., gap-filling, temporal aggregation) and energy balance closure assessment and correction [@wilson2002]. It is focused towards estimating daily and monthly ET from input flux data at half-hourly or hourly frequency, such as from the AmeriFlux dataset. 
+The ``flux-data-qaqc`` open-source and object-oriented Python package is intended for a wide audience including anyone who needs to process EC flux data to estimate daily or monthly ET that has been corrected for energy balance closure. It is also useful for those who wish to perform visual or flag-based quality control and quality assurance (QA/QC) checks on EC flux or meteorological data to filter out or replace any suspect data points before applying the data for any given use. For example, atmospheric and hydrologic modeling and field scientists may find this software useful for simplifying the post-processing of 30-minute EC flux data into daily ET that has been gap-filled and corrected (or not) for energy imbalance using different approaches. The software will also make calculations of reference ET from EC meteorological data and download reference ET from gridded products [@abatzoglou2013]. Output daily and monthly ET estimates from ``flux-data-qaqc`` can then be used by scientists for model validation and applied science, while knowing that the post-processing steps used to generate them are tested and well documented. 
+
+Post-processing steps of EC data commonly involve at least some of the following steps: data formatting, unit conversions, gap-filling, time aggregation, ET and other atmospheric calculations, energy imbalance assessment and correction, and data QA/QC (automatic or visual). Because ``flux-data-qaqc`` provides a framework for these common EC post-processing steps using Python objects, it also allows scientists or engineers to document and share their processing decisions and analyses with others via the Python scripts they write, which is critical in any serious study. Although the tools provided by ``flux-data-qaqc`` are well documented and "high-level", minimal proficiency in Python is required. 
+
+The interactive  time series and scatter plots (e.g. tools for hover, pan, zoom, selection, and linked axes) produced by ``flux-data-qaqa`` are useful for data QA/QC and for anyone interested in understanding or visually documenting the atmospheric and/or hydrologic conditions at an EC tower. Within the same plot document, daily and monthly surface energy balance variables (average fluxes) as well as all commonly measured meteorological variables (e.g. wind speed, temperature, precipitation, vapor pressure deficit, soil moisture) are displayed. In addition, the plots display computed results useful for EC data QA/QC and many ET-oriented analyses, these results include: energy balance closure (daily and monthly), the energy balance ratio initially and after despiking and smoothing, ET as a fraction of gridded reference ET, as well as daily ET gap information. 
+
+Ultimately, obtaining final ET estimates from EC data involves multiple human-driven processing decisions and this presents an uncertainty issue when comparing data that are processed differently. ``flux-data-qaqc`` does not address issues regarding the deployment configuration  of the EC system or the initial processing of high frequency EC data, for which several software exist. Instead, this software provides a framework and provenance for EC data post-processing (e.g., gap-filling, temporal aggregation), energy balance closure analyses, and ET estimates [@wilson2002].
 
[…]
