+++ paper/paper.bib
@@ -165,7 +165,7 @@ abstract = {The energy balance at most surface-atmosphere flux research sites re
 	abstract = {The FLUXNET2015 dataset provides ecosystem-scale data on CO2, water, and energy exchange between the biosphere and the atmosphere, and other meteorological and biological measurements, from 212 sites around the globe (over 1500 site-years, up to and including year 2014). These sites, independently managed and operated, voluntarily contributed their data to create global datasets. Data were quality controlled and processed using uniform methods, to improve consistency and intercomparability across sites. The dataset is already being used in a number of applications, including ecophysiology studies, remote sensing studies, and development of ecosystem and Earth system models. FLUXNET2015 includes derived-data products, such as gap-filled time series, ecosystem respiration and photosynthetic uptake estimates, estimation of uncertainties, and metadata about the measurements, presented for the first time in this paper. In addition, 206 of these sites are for the first time distributed under a Creative Commons (CC-BY 4.0) license. This paper details this enhanced dataset and the processing methods, now made available as open-source codes, making the dataset more accessible, transparent, and reproducible.},
 	number = {1},
 	journal = {Scientific Data},
-	author = {Pastorello, Gilberto and Trotta, Carlo and Canfora, Eleonora and Chu, Housen and Christianson, Danielle and Cheah, You-Wei and Poindexter, Cristina and Chen, Jiquan and Elbashandy, Abdelrahman and Humphrey, Marty and Isaac, Peter and Polidori, Diego and Ribeca, Alessio and van Ingen, Catharine and Zhang, Leiming and Amiro, Brian and Ammann, Christof and Arain, M. Altaf and ArdÃ¶, Jonas and Arkebauer, Timothy and Arndt, Stefan K. and Arriga, Nicola and Aubinet, Marc and Aurela, Mika and Baldocchi, Dennis and Barr, Alan and Beamesderfer, Eric and Marchesini, Luca Belelli and Bergeron, Onil and Beringer, Jason and Bernhofer, Christian and Berveiller, Daniel and Billesbach, Dave and Black, Thomas Andrew and Blanken, Peter D. and Bohrer, Gil and Boike, Julia and Bolstad, Paul V. and Bonal, Damien and Bonnefond, Jean-Marc and Bowling, David R. and Bracho, Rosvel and Brodeur, Jason and BrÃ¼mmer, Christian and Buchmann, Nina and Burban, Benoit and Burns, Sean P. and Buysse, Pauline and Cale, Peter and Cavagna, Mauro and Cellier, Pierre and Chen, Shiping and Chini, Isaac and Christensen, Torben R. and Cleverly, James and Collalti, Alessio and Consalvo, Claudia and Cook, Bruce D. and Cook, David and Coursolle, Carole and Cremonese, Edoardo and Curtis, Peter S. and Dâ€™Andrea, Ettore and da Rocha, Humberto and Dai, Xiaoqin and Davis, Kenneth J. and De Cinti, Bruno and de Grandcourt, Agnes and De Ligne, Anne and De Oliveira, Raimundo C. and Delpierre, Nicolas and Desai, Ankur R.
[…]
