### 2021-10-11 f34781f
remove all caps titles in bibliography
