+++ paper/paper.bib
@@ -20,7 +20,7 @@ DOI = {10.5194/bg-15-5015-2018}
  number = {9},
  pages = {1791--1812},
  publisher = {American Meteorological Society},
- title = {THE GRAPE REMOTE SENSING ATMOSPHERIC PROFILE AND EVAPOTRANSPIRATION EXPERIMENT},
+ title = {The Grape Remote Sensing Atmospheric Profile and Evapotranspiration Experiment},
  volume = {99},
  year = {2018}
 }
@@ -42,7 +42,7 @@ abstract = {The ‘energy imbalance problem’ in micrometeorology arises becaus
 
 @article{foken2008,
 author = {Foken, Thomas},
-title = {THE ENERGY BALANCE CLOSURE PROBLEM: AN OVERVIEW},
+title = {The Energy Balance Closure Problem: An Overview},
 journal = {Ecological Applications},
 volume = {18},
 number = {6},
@@ -90,7 +90,6 @@ note = {also available at \url{https://www.mesonet.org/images/site/ASCE_Evapotra
       url = "https://journals.ametsoc.org/view/journals/bams/82/11/1520-0477_2001_082_2415_fantts_2_3_co_2.xml"
 }
 
-Forrest Melton1,2, Justin Huntington3, Robyn Grimm4, Jamie Herring5, Maurice Hall4, Dana Rollison4, Tyler Erickson6, Richard Allen7, Martha Anderson8, Joshua B. Fisher9,10, Ayse Kilic11,  Gabriel Senay12,  John Volk3, Christopher Hain13, Lee Johnson1,2, Anderson Ruhoff14, Philip Blankenau15, Matt Bromley3, Will Carrara1,2, Britta Daudert3, Conor Doherty16, Christian Dunkerly3, MacKenzie Friedrichs12, Alberto Guzman1.2, Gregory Halverson9, Jody Hansen3, Jordan Harding5, Yanghui Kang17, David Ketchum18, Blake Minor3, Charles Morton3, Samuel Ortega-Salazar12, Thomas Ott3, Mutlu Ozdogan17, Peter Revelle7,11, Mitch Schull8,19, Carlos Wang20, Yun Yang8,19, Ray Anderson8
 
 @article{melton2021,
 author = {Melton, F. and Huntington, J. and Grimm, R. and Herring, J. and Hall, M. and Rollison, D. and Erickson, T. and Allen, R. and Anderson, M. and Fisher, J. B. and Kilic, A. and Senay, G. and Volk, J. and Hain, C. and Johnson, L. and Ruhoff, A. and Blankenau, P. and Bromley, M. and Carrara, W. and Daudert, B. and Doherty, C. and Dunkerly, C. and Friedrichs, M. and Guzman, A. and Halverson, G. and Hansen, J. and Harding, J. and Kang, Y. and Ketchum, D. and Minor, B. and Morton, C. and Ortega-Salazar, S. Ott, T. and Ozdogan, M. and Revelle, P. and Schull, M. and Wang, C. and Yang, Y. and Anderson, R.},
