### 2021-10-11 #3418 @JohnVolk
Thank you @ashwinvis, I just updated those all caps titles in the bibliography. Not sure about incorrect names in the Pastorello et al. (2020), there are special characters that the bibtex doesn't handle well but those names do not show up in the preprint, but I might be missing something.
