### 2021-10-14 #3418 @JohnVolk
Hi @pdebuyl thanks, I don't see any issues in the actual bibliography, the encoding issues are for special characters in certain author names for the Pastorello et al., 2020 citation but those authors are not included in the list that show up in the bibliography because the author list is so long they get cut out. Sorry for not clarifying this the other day when I tried to clean it up, I only fixed the authors that show up in the actual bibliography. Looks like you also removed abstracts and an unused citation, thanks again for doing that.
