### 2021-10-15 #3418 @JohnVolk
Hi @pdebuyl I just updated the title on Zenodo and added all the project members as contributors.

Thanks for the feedback on the paper, happy to hear that and I really appreciate it.

As for the tutorial, I will add a note on that. Although this might have been resolved when I recently added the entire example directory to the PyPI package so if installing with PIP it should come through automatically now.

Thanks again for your work on getting this through and improving the paper.
