### 2021-10-15 #3418 @pdebuyl
PS: now, one of the editors-in-chief of JOSS will take over.
